Modified the buy and sell operations to handle the 'portfolioId' and "shareId" in requests and database.

+++ models/Share.js
@@ -3,7 +3,7 @@ const { DataTypes } = require('sequelize');
 module.exports = (sequelize) => {
   const Share = sequelize.define('Share', {
     symbol: { type: DataTypes.STRING, unique: true },
-    price: DataTypes.DECIMAL(10, 2),
+    price: DataTypes.FLOAT,
   });
 
   Share.associate = (models) => {
